Move ManageEntity one folder down
Updating the namespace change in the process thoughtout the application
